Transition Diagram For Relational Operator in C++
We begin in state 0, the start state.
If we see < as the first
input symbol, then among the lexemes that match the pattern for relop
we can only be looking at <, <>, or <=.
We therefore go to state 1, and look at the next character.
If it is =, then we recognize lexeme <=, enter state 2, and return the
token relop with attribute LE, the symbolic constant representing this
particular comparison operator.

If in state 1 the next instead we have lexeme <>, and enter state 3 to
return an indication that the not-equals operator has been found.

On any other character, the lexeme is <, and we enter state 4 to
return that information.

Note, however, that state 4 has a ∗ to indicate that we must retract
the input one position.

On the other hand, if in state 0 the first character we see is =, then
this one character must be the lexeme.

We immediately return that fact from state 5.

The remaining possibility is that the first character is >.

Then, we must enter state 6 and decide, on the basis of the next
character, whether the lexeme is >= (if we next see the = sign), or
just > (on any other character).

Note that if, in state 0, we see any character besides <, =, or >, we
can not possibly be seeing a relop lexeme, so this transition diagram
will not be used.

Note that if, in state 0, we see any character besides <, =, or >, we
can not possibly be seeing a relop lexeme, so this transition diagram
will not be used.

Note that if, in state 0, we see any character besides <, =, or >, we
can not possibly be seeing a relop lexeme, so this transition diagram
will not be used.
Note that if, in state 0, we see any character besides <, =, or >, we
can not possibly be seeing a relop lexeme, so this transition diagram
will not be used.